# apps/macos/Sources/VioletProtocolClient/openapi.yaml
openapi: 3.1.1
info:
  title: Violet Core API
  version: 0.1.0
  description: Versioned control-plane API for the single-user Violet instance.
servers:
  - url: http://127.0.0.1:4310
paths:
...
  /v1/status:
    get:
      operationId: getCoreStatus
      summary: Report whether Core is sealed or ready.
      security:
        - deviceToken: []
      responses:
        '200':
          description: Current Core status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status.schema'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/context/envelopes:
    post:
      operationId: submitContextEnvelope
      summary: Validate and resolve one explicitly authorized context envelope.
      security:
        - deviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/context-envelope.schema'
      responses:
        '200':
          description: The context is available to the active conversation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/context-receipt.schema'
        '400':
          description: The context envelope is invalid or violates local-policy assertions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.schema'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '410':
          description: The context envelope has expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.schema'
        '423':
          description: Core is sealed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.schema'
  /v1/context/sessions/{sessionId}:
    delete:
      operationId: deleteContextSession
      summary: End a context session and discard all ephemeral evidence.
      security:
        - deviceToken: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The context session was removed or had already expired.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Core is sealed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.schema'
  /v1/chat/stream:
    post:
      operationId: streamChat
      summary: Stream one text interaction as newline-delimited protocol events.
      security:
        - deviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chat-request.schema'
      responses:
        '200':
          description: Newline-delimited ChatStreamEvent values.
          content:
            application/x-ndjson:
              schema:
                type: string
              x-ndjson-schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                $id: https://violet.local/schemas/v1/chat-stream-event.schema.json
                title: ChatStreamEvent
                oneOf:
                  - type: object
                    additionalProperties: false
                    required:
                      - eventId
                      - requestId
                      - type
                    properties:
                      eventId:
                        type: string
                        format: uuid
                      requestId:
                        type: string
                        format: uuid
                      type:
                        const: start
                  - type: object
                    additionalProperties: false
                    required:
                      - content
                      - eventId
                      - requestId
                      - type
                    properties:
                      content:
                        type: string
                        minLength: 1
                      eventId:
                        type: string
                        format: uuid
                      requestId:
                        type: string
                        format: uuid
                      type:
                        const: delta
                  - type: object
                    additionalProperties: false
                    required:
                      - eventId
                      - messageId
                      - requestId
                      - type
                      - usage
                    properties:
                      eventId:
                        type: string
                        format: uuid
                      messageId:
                        type: string
                        format: uuid
                      requestId:
                        type: string
                        format: uuid
                      type:
                        const: complete
                      usage:
                        type: object
                        additionalProperties: false
                        required:
                          - inputTokens
                          - outputTokens
                        properties:
                          inputTokens:
                            type: integer
                            minimum: 0
                          outputTokens:
                            type: integer
                            minimum: 0
                  - type: object
                    additionalProperties: false
                    required:
                      - error
                      - eventId
                      - requestId
                      - type
                    properties:
                      error:
                        $ref: '#/components/schemas/error.schema'
                      eventId:
                        type: string
                        format: uuid
                      requestId:
                        type: string
                        format: uuid
                      type:
                        const: error
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Core is sealed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.schema'
...
components:
  schemas:
...
    status.schema:
      $schema: https://json-schema.org/draft/2020-12/schema
      $id: https://violet.local/schemas/v1/status.schema.json
      title: CoreStatus
      type: object
      additionalProperties: false
      required:
        - service
        - state
        - time
        - version
      properties:
        service:
          const: violet-core
        state:
          type: string
          enum:
            - sealed
            - ready
        reason:
          type: string
          minLength: 1
        time:
          type: string
          format: date-time
        version:
          type: string
          minLength: 1
    error.schema:
      $schema: https://json-schema.org/draft/2020-12/schema
      $id: https://violet.local/schemas/v1/error.schema.json
      title: ApiError
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
        - retryable
      properties:
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]+$
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          format: uuid
        retryable:
          type: boolean
    probability:
      type: number
      minimum: 0
      maximum: 1
    uuid:
      type: string
      format: uuid
    bundleId:
      type: string
      minLength: 3
      maxLength: 255
      pattern: ^[A-Za-z0-9][A-Za-z0-9.-]+$
    normalizedPoint:
      type: object
      additionalProperties: false
      required:
        - x
        - 'y'
      properties:
        x:
          $ref: '#/components/schemas/probability'
        'y':
          $ref: '#/components/schemas/probability'
    image:
      type: object
      additionalProperties: false
      required:
        - data
        - height
        - mediaType
        - sha256
        - width
      properties:
        data:
          type: string
          minLength: 4
          maxLength: 11184812
          pattern: ^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$
        height:
          type: integer
          minimum: 1
          maximum: 8192
        mediaType:
          type: string
          enum:
            - image/jpeg
            - image/png
        sha256:
          type: string
          pattern: ^[a-f0-9]{64}$
        width:
          type: integer
          minimum: 1
          maximum: 8192
    normalizedRect:
      type: object
      additionalProperties: false
      required:
        - height
        - width
        - x
        - 'y'
      properties:
        height:
          $ref: '#/components/schemas/probability'
        width:
          $ref: '#/components/schemas/probability'
        x:
          $ref: '#/components/schemas/probability'
        'y':
          $ref: '#/components/schemas/probability'
    context-envelope.schema:
      $schema: https://json-schema.org/draft/2020-12/schema
      $id: https://violet.local/schemas/v1/context-envelope.schema.json
      title: ContextEnvelope
      type: object
      additionalProperties: false
      required:
        - authorization
        - capturedAt
        - completeness
        - confidence
        - eventId
        - expiresAt
        - payload
        - protocolVersion
        - redactions
        - sensitivity
        - sequence
        - sessionId
        - source
      properties:
        authorization:
          type: object
          additionalProperties: false
          required:
            - controlledSensitiveAllowed
            - grantId
            - mode
            - purpose
            - retention
          properties:
            controlledSensitiveAllowed:
              type: boolean
            grantId:
              $ref: '#/components/schemas/uuid'
            mode:
              const: explicit
            purpose:
              const: conversation
            retention:
              const: ephemeral
        capturedAt:
          type: string
          format: date-time
        completeness:
          $ref: '#/components/schemas/probability'
        confidence:
          $ref: '#/components/schemas/probability'
        eventId:
          $ref: '#/components/schemas/uuid'
        expiresAt:
          type: string
          format: date-time
        payload:
          oneOf:
            - type: object
              additionalProperties: false
              required:
                - text
                - type
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 65536
                type:
                  const: focus.text
            - type: object
              additionalProperties: false
              required:
                - appBundleId
                - type
              properties:
                appBundleId:
                  $ref: '#/components/schemas/bundleId'
                appName:
                  type: string
                  minLength: 1
                  maxLength: 256
                type:
                  const: app.state
            - type: object
              additionalProperties: false
              required:
                - image
                - type
              properties:
                focusPoint:
                  $ref: '#/components/schemas/normalizedPoint'
                image:
                  $ref: '#/components/schemas/image'
                localText:
                  type: string
                  maxLength: 65536
                type:
                  const: screen.snapshot
            - type: object
              additionalProperties: false
              required:
                - image
                - region
                - type
              properties:
                focusPoint:
                  $ref: '#/components/schemas/normalizedPoint'
                image:
                  $ref: '#/components/schemas/image'
                localText:
                  type: string
                  maxLength: 65536
                region:
                  $ref: '#/components/schemas/normalizedRect'
                type:
                  const: focus.region
            - type: object
              additionalProperties: false
              required:
                - transcript
                - type
              properties:
                transcript:
                  type: string
                  minLength: 1
                  maxLength: 65536
                type:
                  const: audio.utterance
        previousEventId:
          $ref: '#/components/schemas/uuid'
        protocolVersion:
          const: '1'
        redactions:
          type: array
          maxItems: 128
          items:
            type: object
            additionalProperties: false
            required:
              - category
              - count
            properties:
              category:
                type: string
                enum:
                  - absolute_secret
                  - controlled_sensitive
                  - secure_field
              count:
                type: integer
                minimum: 1
        sensitivity:
          type: string
          enum:
            - public
            - personal
            - controlled
        sequence:
          type: integer
          minimum: 1
        sessionId:
          $ref: '#/components/schemas/uuid'
        source:
          type: object
          additionalProperties: false
          required:
            - deviceId
            - modality
          properties:
            appBundleId:
              $ref: '#/components/schemas/bundleId'
            deviceId:
              $ref: '#/components/schemas/uuid'
            modality:
              type: string
              enum:
                - accessibility
                - audio
                - screen
      $defs:
        bundleId:
          type: string
          minLength: 3
          maxLength: 255
          pattern: ^[A-Za-z0-9][A-Za-z0-9.-]+$
        image:
          type: object
          additionalProperties: false
          required:
            - data
            - height
            - mediaType
            - sha256
            - width
          properties:
            data:
              type: string
              minLength: 4
              maxLength: 11184812
              pattern: ^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$
            height:
              type: integer
              minimum: 1
              maximum: 8192
            mediaType:
              type: string
              enum:
                - image/jpeg
                - image/png
            sha256:
              type: string
              pattern: ^[a-f0-9]{64}$
            width:
              type: integer
              minimum: 1
              maximum: 8192
        normalizedRect:
          type: object
          additionalProperties: false
          required:
            - height
            - width
            - x
            - 'y'
          properties:
            height:
              $ref: '#/components/schemas/probability'
            width:
              $ref: '#/components/schemas/probability'
            x:
              $ref: '#/components/schemas/probability'
            'y':
              $ref: '#/components/schemas/probability'
        normalizedPoint:
          type: object
          additionalProperties: false
          required:
            - x
            - 'y'
          properties:
            x:
              $ref: '#/components/schemas/probability'
            'y':
              $ref: '#/components/schemas/probability'
        probability:
          type: number
          minimum: 0
          maximum: 1
        uuid:
          type: string
          format: uuid
    context-receipt.schema:
      $schema: https://json-schema.org/draft/2020-12/schema
      $id: https://violet.local/schemas/v1/context-receipt.schema.json
      title: ContextReceipt
      type: object
      additionalProperties: false
      required:
        - acceptedAt
        - eventId
        - expiresAt
        - sessionId
        - status
      properties:
        acceptedAt:
          type: string
          format: date-time
        eventId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        sessionId:
          type: string
          format: uuid
        status:
          const: ready
    chat-request.schema:
      $schema: https://json-schema.org/draft/2020-12/schema
      $id: https://violet.local/schemas/v1/chat-request.schema.json
      title: ChatRequest
      type: object
      additionalProperties: false
      required:
        - message
        - requestId
      properties:
        contextSessionId:
          type: string
          format: uuid
        message:
          type: string
          minLength: 1
          maxLength: 65536
        requestId:
          type: string
          format: uuid
...
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-device-token
  responses:
    Unauthorized:
      description: Device token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error.schema'